refactor: 로또 구매 관련 기능들 packaging
