+++ src/main/kotlin/lotto/domain/shop/LottoGame.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.shop
 
 data class LottoGame(
     val numbers: List<Int>,
