+++ src/main/kotlin/lotto/domain/shop/LottoPurchaseResult.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.shop
 
 data class LottoPurchaseResult(
     val lottoGames: List<LottoGame>,
