+++ src/main/kotlin/lotto/domain/shop/LottoShop.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.shop
 
 interface LottoShop {
