+++ src/main/kotlin/lotto/domain/shop/RealLottoShop.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.shop
 
 class RealLottoShop : LottoShop {
